Migrated CollectionEditor to JavaFX, fixed issues ZipBuilder and CacheManager

+++ desktop/src/main/java/attilathehun/songbook/environment/Installer.java
@@ -1,6 +1,6 @@
 package attilathehun.songbook.environment;
 
-import attilathehun.songbook.util.ZipUtil;
+import attilathehun.songbook.util.ZipBuilder;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -48,8 +48,10 @@ public class Installer {
                 IS_TEMP_RESOURCES_ZIP_FILE = true;
                 downloadRemoteFile(REMOTE_RESOURCES_ZIP_FILE, "resources.zip");
             }
-            new ZipUtil().extractZip(RESOURCES_ZIP_FILE_PATH, Environment.getInstance().settings.environment.RESOURCE_FILE_PATH);
-            new File(RESOURCES_ZIP_FILE_PATH).delete();
+            ZipBuilder.extract(RESOURCES_ZIP_FILE_PATH, Environment.getInstance().settings.environment.RESOURCE_FILE_PATH);
+            if (IS_TEMP_RESOURCES_ZIP_FILE) {
+                new File(RESOURCES_ZIP_FILE_PATH).delete();
+            }
             logger.info("Finished installing resources");
         } catch (IOException e) {
             logger.error(e.getMessage(), e);
@@ -68,7 +70,9 @@ public class Installer {
                 IS_TEMP_SCRIPTS_ZIP_FILE = true;
                 downloadRemoteFile(REMOTE_SCRIPTS_ZIP_FILE, "scripts.zip");
             }
-            new ZipUtil().extractZip(SCRIPTS_ZIP_FILE_PATH, Environment.getInstance().settings.environment.SCRIPTS_FILE_PATH);
+            if (IS_TEMP_SCRIPTS_ZIP_FILE) {
+                ZipBuilder.extract(SCRIPTS_ZIP_FILE_PATH, Environment.getInstance().settings.environment.SCRIPTS_FILE_PATH);
+            }
             logger.info("Finished installing scripts");
         } catch (IOException e) {
             logger.error(e.getMessage(), e);
